requirements.txt
python-binance
pandas
matplotlib
python-dotenv
